+++ lib/nretnil-fake-data.rb
@@ -1,6 +1,8 @@
-# Base Class
+require 'rubygems'
+require 'date'
+
 require 'nretnil/fake-data/colors'
-require 'nretnil/fake-data/data'
+require 'nretnil/fake-data/fake_data'
 require 'nretnil/fake-data/names'
 require 'nretnil/fake-data/version'
 require 'nretnil/fake-data/words'